Add the controlsfx-8.0.6_20.jar as a direct download
The reason is that in controlsfx-8.0.7 and above the Dialogs API is marked as deprecated because it will be moved
to the official JDK 8u40. JDK 8u40 is not released yet and I don't want to change all tutorials to depend on JDK 8u40.
The controlsfx-8.0.6_20.jar is perfectly fine and doesn't confuse new users with a deprecated flag.