Changed Leafs to Leaves
In the comments, someone requested to changed it to leaves and the owner responded by making a pull request to change it. I am not the person who asked, but I would also greatly appreciate if this was fixed, and so I am here, making my first pull request into a repo I don't own.